// cypress/support/types.ts
enum UserStatus {
  Active,
}

export enum Gender {
  Male = "Male",
  Female = "Female",
}

export enum StartInvestmentTimeFrame {
  ThisMonth = "ThisMonth",
  Next3Months = "Next3Months",
  Next6Months = "Next6Months",
  After1Year = "After1Year",
}

export enum selfClassification {
  ProfessionalInvestor = "ProfessionalInvestor",
  PrivateInvestor = "ProfessionalPrivateInvestor",
  InstitutionalInvestor = "InstitutionalInvestor",
  NoneOfTheAbove = "None",
}


export type User = {
  userId: string
  email: string
  emailVerified: boolean
  picture: string
  nickname: string
  status: UserStatus
  profile: Profile
  name?: string
  mandateId?: string
  isFirstTimeLogin?: boolean
  user_metadata?: {
    phone_number?: string
  }
  roles: string[]
  phoneNumberVerified?: boolean
  contactId?: string
}

export type InvestorSurvey = {
  investMinimumAmount?: string
  whenToStartInvestment?: string
  interestedInvestments?: InterestedInvestments[]
  investmentExperience?: string
  priorInvExperience?: PriorInvestment[]
  otherInterestedInvDetails?: string
  otherPriorDetails?: string
}

export type RelationshipManager = {
  assigned: boolean
  manager?: {
    firstName: string
    lastName: string
    email: string
  }
}

export type ClientRelationshipManager = {
  email?: string
  manager: {
    name: string
    email?: string
  }
}

export type UserStatuses = {
  status: UserQualificationStatus
}

export type Profile = {
  firstName?: string
  lastName?: string
  countryOfResidence?: any
  nationality?: any
  region?: any
  phoneCountryCode?: string
  phoneNumber?: string
  // Note: Added boolean support as a temporary backend workaround.
  // We should look at creating a separate API client specifically for form types on the FE.
  // For example, `/services/client` and `/services/server`.
  isTaxableInUS?: YesOrNo | boolean
  isDefinedSophisticatedByCMA?: YesOrNo | boolean
  isAccreditedByCBB?: YesOrNo | boolean
  relationshipManager?: RelationshipManager
  investorSurvey?: InvestorSurvey
  preProposalInitialAction?: string
  flagName?: string
  selfClassification?: selfClassification
  investmentFlow?: ProfileInvestmentFlow | null
}

export enum ProfileInvestmentFlow {
  Advisory = "Advisory",
  Solutions = "Solutions",
}
// ...
export type YesOrNo = "yes" | "no"